45 react query/6 infinite scroll

+++ src/features/ticket/actions/delete-ticket.ts
@@ -1,20 +1,33 @@
 "use server";
+import { setCookieByKey } from "@/actions/cookies";
+import {
+  fromErrorToActionState,
+  toActionState,
+} from "@/components/form/util/to-action-state";
+import { getAuthOrRedirect } from "@/features/auth/queries/get-auth-or-redirect";
+import { isOwner } from "@/features/auth/utils/is-owner";
 import { prisma } from "@/lib/prisma";
 import { ticketsPath } from "@/paths";
 import { revalidatePath } from "next/cache";
 import { redirect } from "next/navigation";
-import { setCookieByKey } from "@/actions/cookies";
-import { formErrorToActionState } from "@/components/form/util/to-action-state";
 
 export const deleteTicket = async (id: string) => {
+  const { user } = await getAuthOrRedirect();
   try {
+    const ticket = await prisma.ticket.findUnique({
+      where: { id },
+    });
+
+    if (!ticket || !isOwner(user, ticket)) {
+      return toActionState("ERROR", "Not Authorized");
+    }
     await prisma.ticket.delete({
       where: {
         id,
       },
     });
   } catch (error) {
-    return formErrorToActionState(error);
+    return fromErrorToActionState(error);
   }
 
   revalidatePath(ticketsPath());
